requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-timeout==4.0.3
attrs==21.4.0
backcall==0.2.0
backoff==2.2.1
backports.tarfile==1.2.0
beautifulsoup4==4.12.3
bleach==4.1.0
blinker==1.9.0
build==1.2.2.post1
CacheControl==0.14.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cleo==2.1.0
click==8.1.7
crashtest==0.4.1
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.9
distro==1.9.0
dulwich==0.22.8
entrypoints==0.4
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==0.8.3
fastapi==0.116.1
fastjsonschema==2.21.1
filelock==3.17.0
findpython==0.6.2
Flask==3.1.0
fonttools==4.29.1
fpdf==1.7.2
frozenlist==1.5.0
googleapis-common-protos==1.70.0
greenlet==3.1.1
h11==0.14.0
HTMLParser==0.0.2
httpcore==1.0.6
httpx==0.27.2
idna==3.10
importlib_metadata==8.5.0
installer==0.7.0
ipykernel==6.9.1
ipython==8.1.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.18.1
Jinja2==3.1.4
jiter==0.7.1
joblib==1.4.2
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keyring==25.6.0
kiwisolver==1.3.2
langfuse==3.3.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.5.1
matplotlib-inline==0.1.3
mdurl==0.1.2
mistune==0.8.4
more-itertools==10.6.0
msgpack==1.1.0
multidict==6.1.0
nbclient==0.5.12
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.8
numpy==1.22.3
openai==0.28.0
openpyxl==3.1.5
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-http==1.36.0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
packaging==24.2
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
pbs-installer==2025.2.12
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pkginfo==1.12.1.2
platformdirs==4.3.6
playwright==1.48.0
poetry==2.1.1
poetry-core==2.1.1
prometheus-client==0.13.1
prompt-toolkit==3.0.28
propcache==0.2.0
protobuf==6.32.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==2.9.2
pydantic_core==2.23.4
pyee==12.0.0
Pygments==2.18.0
pyparsing==3.0.7
pyproject_hooks==1.2.0
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2021.3
PyYAML==6.0.2
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.1
RapidFuzz==3.12.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
scikit-learn==1.5.2
scipy==1.8.0
seaborn==0.11.2
selectolax==0.3.26
Send2Trash==1.8.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
stack-data==0.2.0
starlette==0.47.3
terminado==0.13.3
testpath==0.6.0
threadpoolctl==3.5.0
tomli==2.2.1
tomlkit==0.13.2
tornado==6.1
tqdm==4.67.0
traitlets==5.1.1
trove-classifiers==2025.3.3.18
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.35.0
virtualenv==20.29.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==3.1.3
widgetsnbextension==3.5.2
wrapt==1.17.3
xattr==1.1.4
yarl==1.17.1
zipp==3.21.0
zstandard==0.23.0
